Global key/store view
Pull all (user) keys local to get a global view.
Locking does frequent remote-key flush, so in a large cluster most
user-keys are flushed from local views during the locking.  This
restores them.
Fix an NPE in Frame.subFrame.

+++ src/test/java/hex/gbm/GBMTest.java
@@ -254,7 +254,7 @@ public class GBMTest extends TestUtil {
       if( model != null ) model.delete();
 
     } finally {
-      gbm.source.delete();      // Remove original hex frame key
+      if( gbm.source != null ) gbm.source.delete(); // Remove original hex frame key
       gbm.remove();             // Remove GBM Job
     }
   }
